feat: possibility to delete lists

+++ components/ListItem.tsx
@@ -3,58 +3,40 @@ import { deleteList } from "@/db/db";
 import type { List } from "@/types";
 import Feather from "@expo/vector-icons/Feather";
 import { Link } from "expo-router";
-import { StyleSheet, Text } from "react-native";
-import ReanimatedSwipeable from 'react-native-gesture-handler/ReanimatedSwipeable';
+import { StyleSheet, Text, View } from "react-native";
 import Animated, { LinearTransition, useAnimatedStyle } from "react-native-reanimated";
+import { Alert } from "react-native";
 
 export default function ListItem({ data_creazione, id }: List) {
     const handleDelete = async () => {
         if (!id) return;
         try {
-            await deleteList(id);
+            Alert.alert("Conferma eliminazione", "Sei sicuro di voler eliminare questa lista?", [
+                {
+                    text: "Annulla", style: "cancel"
+                },
+                {
+                    text: "Elimina",
+                    onPress: async () => {
+                        await deleteList(id);
+                    },
+                    style: "destructive"
+                }
+            ]);
         } catch (error) {
             console.error("deleteList failed:", error);
         }
     };
 
-    function leftSwipeAction(progress: any, dragValue: any) {
-        const styleAnimation = useAnimatedStyle(() => ({
-            opacity: progress.value,
-            width: progress.value * 30,
-        }));
-
-        return (
-            <Animated.View style={[styles.deleteLeft, styleAnimation]}>
-                <Feather name="trash-2" size={24} color={"#b30000"} />
-            </Animated.View>
-        );
-    }
-
-    function rightSwipeAction(progress: any, dragValue: any) {
-        const styleAnimation = useAnimatedStyle(() => ({
-            opacity: progress.value,
-            width: progress.value * 30,
-        }));
-
-        return (
-            <Animated.View style={[styles.deleteRight, styleAnimation]}>
-                <Feather name="trash-2" size={24} color={"#b30000"} />
-            </Animated.View>
-        );
-    }
-
     return (
-        <ReanimatedSwipeable
-            renderLeftActions={leftSwipeAction}
-            renderRightActions={rightSwipeAction}
-            onSwipeableOpen={handleDelete}
-        >
             <Animated.View style={styles.wrapper} layout={LinearTransition}>
                 <Link style={styles.link} href={{ pathname: "/lists/[id]", params: { id } }}>
                     <Text style={styles.text}>{data_creazione}</Text>
                 </Link>
+               <View style={styles.deleteContainer}>
+                    <Feather name="trash-2" size={24} color={"#fff"} onPress={handleDelete} />
+               </View>
             </Animated.View>
-        </ReanimatedSwipeable>
     );
 }
 
@@ -64,6 +46,8 @@ const styles = StyleSheet.create({
         borderRadius: 12,
         padding: 8,
         marginTop: 10,
+        flexDirection: "row",
+        gap:20
     },
     link: {
         paddingVertical: 6,
@@ -73,22 +57,11 @@ const styles = StyleSheet.create({
         color: "#fff",
         fontFamily: "Quicksand_400Regular",
     },
-    deleteLeft: {
-        backgroundColor: "#ff6467",
+    deleteContainer:{
         justifyContent: "center",
-        alignItems: "flex-start",
-        borderRadius: 12,
-        marginTop: 10,
-        paddingHorizontal: 12,
-        paddingVertical: 8,
-    },
-    deleteRight: {
-        backgroundColor: "#ff6467",
-        justifyContent: "center",
-        alignItems: "flex-end",
-        borderRadius: 12,
-        marginTop: 10,
-        paddingHorizontal: 12,
-        paddingVertical: 8,
-    },
+        alignItems: "center",
+        borderRadius:24,
+        padding:4,
+        backgroundColor:Colors.GIALLO_CHIARO,
+    }
 });
